+++ src/codigo/TP taller/TP taller/Vista/Dibujables/Particula.cpp
@@ -22,6 +22,6 @@ void Particula::actualizar(){
 
 Particula::~Particula(){}
 
-void Particula::dibujar(SDL_Renderer* renderer, int corrimientoX, int corrimientoY, int escalaZoom, int anchoPx, int altoPx){
+void Particula::dibujar(SDL_Renderer* renderer, int corrimientoX, int corrimientoY, float escalaZoom, int anchoPx, int altoPx){
 	DibujableTextura::dibujar(renderer, corrimientoX, corrimientoY, escalaZoom, anchoPx, altoPx);
 }
